♻️ 강재수불부 본을 Double 형식으로 변경

+++ src/main/java/com/lineinc/erp/api/server/domain/steelmanagementv2/entity/SteelManagementDetailV2.java
@@ -105,7 +105,7 @@ public class SteelManagementDetailV2 extends BaseEntity {
      */
     @DiffInclude
     @Column
-    private Integer count;
+    private Double count;
 
     /**
      * 총무게
